# tests/snapshots/openapi_routers.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "guguwebui routers snapshot",
    "version": "test"
  },
  "paths": {
...
    "/api/web-config": {
      "get": {
        "summary": "Api Get Web Config",
        "description": "获取Web配置",
        "operationId": "api_get_web_config___api_web-config__GET",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Api Save Web Config",
        "description": "保存Web配置（结构化字段，无 action 分派；为 None 的字段不修改）",
        "operationId": "api_save_web_config___api_web-config__PUT",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebConfigSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/config-files": {
      "get": {
        "summary": "Api Load Config",
        "description": "加载配置文件，统一返回 {path, type, content, config_data} 文档",
        "operationId": "api_load_config___api_config-files__GET",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "文件路径（受 SafePath 约束）",
              "title": "Path"
            },
            "description": "文件路径（受 SafePath 约束）"
          },
          {
            "name": "translation",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "是否加载 _lang 翻译文件",
              "default": false,
              "title": "Translation"
            },
            "description": "是否加载 _lang 翻译文件"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "解析类型：auto/json/yml/yaml/properties/html",
              "default": "auto",
              "title": "Type"
            },
            "description": "解析类型：auto/json/yml/yaml/properties/html"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Api Save Config",
        "description": "写入配置文件：content=原始文本（原 save_config_file），config_data=结构化（原 save_config）",
        "operationId": "api_save_config___api_config-files__PUT",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "文件路径（受 SafePath 约束）",
              "title": "Path"
            },
            "description": "文件路径（受 SafePath 约束）"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigFileWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/custom-assets/{kind}": {
      "get": {
        "summary": "Get Custom Asset",
        "description": "读取全局自定义文件 overall.css / overall.js",
        "operationId": "get_custom_asset___api_custom-assets_{kind}__GET",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "css",
                "js"
              ],
              "type": "string",
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Save Custom Asset",
        "description": "保存全局自定义文件 overall.css / overall.js",
        "operationId": "save_custom_asset___api_custom-assets_{kind}__PUT",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "css",
                "js"
              ],
              "type": "string",
              "title": "Kind"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomAssetContent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiSuccessEnvelope": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiSuccessEnvelope",
        "description": "统一成功外壳（2.3）：status/message?/data?。"
      },
...
      "ConfigFileWriteRequest": {
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content",
            "description": "原始文本内容（原 save_config_file 语义）"
          },
          "config_data": {
            "title": "Config Data",
            "description": "结构化配置内容（原 save_config 语义）"
          }
        },
        "type": "object",
        "title": "ConfigFileWriteRequest",
        "description": "写入配置文件：content（原始文本）与 config_data（结构化）二选一"
      },
      "CustomAssetContent": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content",
            "description": "自定义文件内容"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "CustomAssetContent"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebConfigSaveRequest": {
        "properties": {
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "port": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port"
          },
          "super_account": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Super Account"
          },
          "disable_admin_login_web": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disable Admin Login Web"
          },
          "enable_temp_login_password": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enable Temp Login Password"
          },
          "ai_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Api Key"
          },
          "ai_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Model"
          },
          "ai_api_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Api Url"
          },
          "mcdr_plugins_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mcdr Plugins Url"
          },
          "repositories": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repositories"
          },
          "ssl_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssl Enabled"
          },
          "ssl_certfile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssl Certfile"
          },
          "ssl_keyfile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssl Keyfile"
          },
          "ssl_keyfile_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssl Keyfile Password"
          },
          "public_chat_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Chat Enabled"
          },
          "public_chat_to_game_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Chat To Game Enabled"
          },
          "chat_verification_expire_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Verification Expire Minutes"
          },
          "chat_session_expire_hours": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Session Expire Hours"
          },
          "force_standalone": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Force Standalone"
          },
          "log_capture_compat_mode": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Capture Compat Mode"
          },
          "icp_records": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icp Records"
          },
          "panel_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Panel Role"
          },
          "panel_slaves": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Panel Slaves"
          },
          "panel_master": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Panel Master"
          }
        },
        "type": "object",
        "title": "WebConfigSaveRequest",
        "description": "Web 配置保存（REST 结构化字段，无 action 分派；为 None 的字段不修改）"
      },
...
    }
  }
}